Create README.md
In this project, I dealt with the problem of forecasting the future values of Time Series of the Wikipedia pages traffic. First I make the series stationary and then apply ARIMA model by estimating the p, d, q parameters of the model using ACF and PACF plots. Later used the LSTM network and ensemble the two to get a better estimate.